renaming to nav2_simple_commander
