--- nav2_python_commander/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/nav2_python_commander
-[install]
-install-scripts=$base/lib/nav2_python_commander
