+++ nav2_simple_commander/nav2_simple_commander/example_nav_through_poses.py
@@ -19,7 +19,7 @@ from geometry_msgs.msg import PoseStamped
 from rclpy.duration import Duration
 import rclpy
 
-from nav2_python_commander.robot_navigator import BasicNavigator, NavigationResult
+from nav2_simple_commander.robot_navigator import BasicNavigator, NavigationResult
 
 '''
 Basic navigation demo to go to poses.
