+++ nav2_simple_commander/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>nav2_python_commander</name>
+  <name>nav2_simple_commander</name>
   <version>1.0.0</version>
   <description>An importable library for writing mobile robot applications in python3</description>
   <maintainer email="stevenmacenski@gmail.com">steve</maintainer>
