+++ nav2_simple_commander/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/nav2_simple_commander
+[install]
+install-scripts=$base/lib/nav2_simple_commander
